backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
anthropic>=0.52.0
httpx==0.27.0
chromadb>=0.5.0
pydantic==2.8.0
python-dotenv==1.0.0
pytest==8.3.0
pytest-asyncio==0.24.0
respx==0.23.1
sse-starlette==2.1.0
pytz>=2024.1
